backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
motor==3.7.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.10.1
pymongo[srv]==4.11.3
pydantic[email]==2.2.0
Authlib==1.5.1
itsdangerous==2.2.0
httpx==0.28.1
python-dateutil==2.9.0.post0
google-genai===1.13.0
python-multipart==0.0.9